utils: Separate context setting reads
The state path utility currently constructs persistent session paths and
also owns the reader for the stored unified-diff context setting.

The mixed responsibility prevents path construction from becoming a
small, auditable interface whose calls only calculate filesystem names.

This commit begins making path construction side-effect free by moving
the context setting reader into a neutral utility module. Consumers now
import the setting from its dedicated owner.

Subsequent commits will make the path getters pure, enforce the boundary
in tests, and document the resulting storage contract.
