+++ src/git_stage_batch/output/candidate_preview.py
@@ -3,8 +3,8 @@
 from __future__ import annotations
 
 from ..batch.operation_candidate_types import OperationCandidatePreview
+from ..utils.context_lines import get_context_lines
 from ..i18n import _
-from ..utils.paths import get_context_lines
 from . import candidate_preview_snippets
 from . import candidate_preview_summary
 from .candidate_preview_commands import (
